feat: Add support for CIVITAI_TOKEN and HF_TOKEN environment variables
Closes #280.

Adds support for the standardized Hugging Face and CivitAI API token environment variables.

This provides a safe and easy way to provide transient secrets without leaking them to disk (nothing gets written to the config).

It also makes it much easier to integrate Comfy-CLI with scripting, since the user no longer has to infinitely provide the "--set-X-token" flags every single time they run the command anymore (which they had to do in the past, just to be sure that the latest token was always used).

The order of priority is `--set-X-token`, then the environment variables (if they exist), and lastly the static config fallback values.

This new feature is now documented in the README and in various error messages.

The code has also been refactored, and the error messages have been improved for clarity and consistency.